Rename: feature-sendtab -> feature-accounts-push
