+++ .buildconfig.yml
@@ -40,6 +40,10 @@ projects:
     path: components/feature/accounts
     description: 'Component for tying an account manager with the tabs feature to facilitate auth flows.'
     publish: true
+  feature-accounts-push:
+    path: components/feature/accounts-push
+    description: 'Feature of use cases for FxA Account that work with push support.'
+    publish: true
   feature-app-links:
     path: components/feature/app-links
     description: 'Component opening URLs in other non-browser apps.'
@@ -71,10 +75,6 @@ projects:
   feature-search:
     path: components/feature/search
     description: 'Feature implementation connecting an engine implementation with the search module.'
-    publish: true
-  feature-sendtab:
-    path: components/feature/sendtab
-    description: 'Feature of use cases for sending tabs between devices in an FxA Account.'
     publish: true
   feature-session:
     path: components/feature/session
