+++ components/feature/accounts-push/src/main/AndroidManifest.xml
@@ -2,4 +2,4 @@
    - License, v. 2.0. If a copy of the MPL was not distributed with this
    - file, You can obtain one at http://mozilla.org/MPL/2.0/. -->
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="mozilla.components.feature.sendtab" />
+    package="mozilla.components.feature.accounts.push" />
