+++ components/feature/accounts-push/src/test/java/mozilla/components/feature/accounts/push/AccountObserverTest.kt
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package mozilla.components.feature.sendtab
+package mozilla.components.feature.accounts.push
 
 import mozilla.components.concept.sync.AuthType
 import mozilla.components.feature.push.AutoPushFeature
